package installs, basic tests done

+++ src/nova_memory/memory/conversations_memory.py
@@ -13,16 +13,19 @@ class ConversationsMemory:
         result = self._conversations_repo.add_conversation(title)
         if result['error'] is not None:
             raise Exception(result['error'])
-        return result['data']
+        self._conversations = self._fetch_conversations()
     
     def delete_conversation(self, id: int) -> None:
+        if id == 1:
+            print("Cannot delete Default Conversation")
+            return
         result = self._conversations_repo.delete_conversation(id)
         if result['error'] is not None:
             raise Exception(result['error'])
         self._conversations = self._fetch_conversations()
     
     def undelete_conversation(self, id: int) -> None:
-        result = self._conversations_repo.add_conversation(id)
+        result = self._conversations_repo.undelete_conversation(id)
         if result['error'] is not None:
             raise Exception(result['error'])
         self._conversations = self._fetch_conversations()
